refactor: split out table component
